Add Debian-Buster-specific deployment steps
On Debian LXC networking is disabled by default and needs to be enabled
either per container or globally. Enable it globally so that it becomes
available to all LXC containers spawned by LAVA.

On Debian Buster, apparmor needs to be installed on the host when using
the Debian LXC template.

Signed-off-by: Karsten Tausche <karsten@fairphone.com>